Split getting started renderer into testable component
