+++ src/vs/workbench/contrib/welcomeGettingStarted/common/gettingStartedContent.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import 'vs/workbench/contrib/welcomeGettingStarted/common/media/example_markdown_media';
+import 'vs/workbench/contrib/welcomeGettingStarted/common/media/theme_picker';
 import 'vs/workbench/contrib/welcomeGettingStarted/common/media/notebookProfile';
 import { localize } from 'vs/nls';
 import { Codicon } from 'vs/base/common/codicons';
@@ -198,7 +198,7 @@ export const walkthroughs: GettingStartedWalkthroughContent = [
 						'onSettingChanged:workbench.colorTheme',
 						'onCommand:workbench.action.selectTheme'
 					],
-					media: { type: 'markdown', path: 'example_markdown_media', }
+					media: { type: 'markdown', path: 'theme_picker', }
 				},
 				{
 					id: 'settingsSync',
@@ -284,7 +284,7 @@ export const walkthroughs: GettingStartedWalkthroughContent = [
 						'onSettingChanged:workbench.colorTheme',
 						'onCommand:workbench.action.selectTheme'
 					],
-					media: { type: 'markdown', path: 'example_markdown_media', }
+					media: { type: 'markdown', path: 'theme_picker', }
 				},
 				{
 					id: 'settingsSyncWeb',
