+++ src/vs/workbench/contrib/welcomeGettingStarted/common/media/theme_picker.ts
